refactor: 지역 select box를 공용 컴포넌트로 분리
